backend/quiz_app_backend/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-filter==1.1.0
djangorestframework==3.15.2
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
setuptools==72.1.0
sqlparse==0.5.1
wheel==0.43.0
